move UI layout API endpoint to ModelViewSetMixin
