+++ changes/3767.added
@@ -1,1 +1,1 @@
-Added `DetailViewConfigMixin` to appropriate ModelViewSets.
+Added `detail-view-config` API endpoint to `ModelViewSetMixin`.
