+++ changes/3767.changed
@@ -1,2 +1,1 @@
 Enhanced error handling for fetcher in React UI `ObjectRetrieveView`.
-Moved `detail-view-config` API endpoint to `DetailViewConfigMixin`.
